+++ CHANGELOG.md
@@ -10,3 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - First release of ftlid
 
+## [v0.1.1] - 2019-11-26
+### Added
+- Possibility to pre-load a model to avoid loading it on every request
+- Default model is also pre-loaded